Fixing undefined behavior when _ has been declared as some value
Typically _ is used in Lua to signify that we are choosing to skip some
parameter. This is fine as long as no other scripts have accidentally
(forgetting to localize) or purposefully set _ to a value, e.g.
_ = true

In this specific case, if _ is true, it will cause construct.SetPhysProp
to horribly break.