Resolved PlayerSpawnProp and PlayerSpawnedProp issues
PlayerSpawnProp returns true if the player is allowed to spawn this
prop, so we should check only if it returns false to block it.

PlayerSpawnedProp doesn't return anything (nil) so we should check if
they return anything other than nil to block it.

Also, PlayerSpawnedProp already calls ply:AddCount, so we no longer need
to use it (otherwise it would make the gamemode think you have twice as
many props spawned out as you do).