Updated the NoCollide option so it actually nocollides all entities with each other
The old behavior was cheap and only nocollided the previous entity with
the current entity -- not a true nocollide since every other entity
would still collide with it.

The new behavior will loop through each entity and nocollide it with
every other entity, so none of them collide with any other entity in the
stack.