Made TOOL:LeftClick fail-fast
I don't particularly know what reasoning I had for making the for-loop continue if one of the prop check hooks failed. If any single one of the props fails a check, it should follow that all of the remaining props to be created in the stack would also fail.

For example, when a player hits the prop limit, this should no longer needlessly create and check props that would fail after the prop limit was initially hit.