Fixing stacker entity count-related bugs
I forgot to add checks that ensured the player entity was valid before trying to decrement their total stacker entities when an entity was removed. If the player has disconnected before the entity was removed, removing it later would cause a relatively harmless, yet annoying error. This should now be resolved.

The other fix addresses the possibility that the TotalStackerEnts field does not get created on the player when they initially spawn. If for some reason it failed, trying to decrement the player's total stacker entity count would cause an error and break things in a bad way. This pull request also addresses this problem.