Fixing a longstanding issue with rotations that didn't treat the new rotations relative to the user's selected direction.
For example, with the new behavior a positive pitch will ALWAYS mean an UPWARD angle RELATIVE to the new direction. Basically, instead of rotation based on the prop's fixed directions, we rotate on the new angle depending on the user's selection. E.g., selecting "Behind" means we want to stack as if the prop's true "Behind" (i.e. -ent:Forward()) is actually it's "Front". Stacking "Up" means the prop's "Up" angle becomes it's new "Front" angle, etc.