Reverting nocollide functionality back to original implementation
Due to the way I had implemented ApplyNoCollide, a large number of
constraints were being created that didn't really need to be. The change
reduces the number of nocollide constraints created from a geometric
series (almost) to just N-1.

Additionally, if clients have both weld and nocollide selected, only a
single weld constraint is created because I'm now utilizing the
nocollide parameter of the weld constraint.

The reason any of this matters is that there is a limit to the number of
constraints that can be out on the server at any given moment. Once it's
hit, any further constraints will break until existings ones are removed
to free up more constraint slots. So, this change reduces the number of
constraints created and helps slow down how quickly the limit is
reached.