Added ability to change decimal precision of offset and rotation sliders
Switched to using DForm:NumSlider to control the number of decimals places for x/y/z offset and pitch/yaw/roll rotation values.

Added stacker_improved_offset_decimals and stacker_improved_rotate_decimals to allow the user to change the precision on the fly, with convar callbacks to rebuild the menu to use the new precision values.